fix: update int64 data type

+++ internal/dto/asesor/create.go
@@ -1,7 +1,7 @@
 package dto
 
 type CreateAsesortReq struct {
-	ClientId int    `json:"client_id" validate:"required"`
+	ClientId int64  `json:"client_id" validate:"required"`
 	Username string `json:"username" validate:"required,username"`
 	Photo    string `json:"photo" validate:"required"`
 	About    string `json:"about" validate:"required"`
